% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.plot.R
\name{.residual.plot}
\alias{.residual.plot}
\title{Residual Plots}
\usage{
.residual.plot(yhat, epsilonhat, tepsilonhat, h)
}
\arguments{
\item{yhat}{Numeric vector of length \code{n} or \code{n} by \code{1} numeric matrix.
\eqn{n \times 1} vector of predicted values of \eqn{\mathbf{y}}
\eqn{\left( \mathbf{\hat{y}} \right)}.}

\item{epsilonhat}{Numeric vector of length \code{n} or \code{n} by 1 numeric matrix.
\eqn{n \times 1} vector of residuals \eqn{\left( \boldsymbol{\hat{\epsilon}} \right)}.}

\item{tepsilonhat}{Numeric vector of length \code{n} or \code{n} by 1 numeric matrix.
\eqn{n \times 1} vector of studentized residuals.}

\item{h}{Numeric vector of length \code{n} or \code{n} by 1 numeric matrix.
\eqn{n \times 1} vector of leverage values.}
}
\description{
Displays a plot with four panels:
\itemize{
\item residuals vs. fitted values
\item normal qq plot of residuals
\item scale location
\item residuals vs. leverage
}
}
\details{
Based on the diagnostic plots in the \code{car} package.
}
\examples{
model <- lm(
  wages ~ gender + race + union + education + experience,
  data = jeksterslabRdatarepo::wages
)
yhat <- as.vector(predict(model))
epsilonhat <- as.vector(residuals(model))
tepsilonhat <- as.vector(rstudent(model))
h <- as.vector(hatvalues(model))
.residual.plot(
  yhat = yhat, tepsilonhat = tepsilonhat,
  epsilonhat = epsilonhat, h = h
)
}
\seealso{
Other plotting functions: 
\code{\link{.scatter.plot}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{plotting functions}
\keyword{plot}
