% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.plot.R
\name{.scatter.plot}
\alias{.scatter.plot}
\title{Scatter Plot Matrix}
\usage{
.scatter.plot(data)
}
\arguments{
\item{data}{Matrix or data frame.}
}
\description{
Adapted
from \url{http://www.sthda.com/english/wiki/scatter-plot-matrices-r-base-graphs} .
}
\examples{
.scatter.plot(data = jeksterslabRdatarepo::wages)
}
\seealso{
Other plotting functions: 
\code{\link{.residual.plot}()}
}
\concept{plotting functions}
\keyword{plot}
